Refactors to reduce bloat
