+++ client/src/components/PasswordForget/PasswordForgetLink.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import { Link } from 'react-router-dom'
+import * as ROUTES from '../../constants/routes'
+
+const PasswordForgetLink = () => (
+  <p>
+    <Link to={ROUTES.PASSWORD_FORGET}>Forgot Password?</Link>
+  </p>
+)
+
+export default PasswordForgetLink
