+++ client/src/components/PasswordForget/PasswordForgetPage.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import { PasswordForgetForm } from './index'
+
+export const PasswordForgetPage = () => (
+  <div>
+    <h1>PasswordForget</h1>
+    <PasswordForgetForm />
+  </div>
+)
