+++ client/src/components/PasswordForget/index.js
@@ -1,79 +1,40 @@
-import React, { Component } from 'react'
-import { Link } from 'react-router-dom'
-
+import React, { useState } from 'react'
 import { withFirebase } from '../../containers/Firebase'
-import * as ROUTES from '../../constants/routes'
 
-const PasswordForgetPage = () => (
-  <div>
-    <h1>PasswordForget</h1>
-    <PasswordForgetForm />
-  </div>
-)
+const PasswordForgetFormBase = props => {
+  const [email, setEmail] = useState('')
+  const [error, setError] = useState(null)
 
-const INITIAL_STATE = {
-  email: '',
-  error: null
+  const onSubmit = async e => {
+    const { firebase } = props
+    const result = await firebase.doPasswordReset(email)
+    try {
+      setEmail(result)
+    } catch (e) {
+      setError(error)
+    }
+
+    e.preventDefault()
+  }
+
+  const onChange = e => {
+    setEmail(e.target.value)
+  }
+
+  return (
+    <form onSubmit={onSubmit}>
+      <input
+        name="email"
+        value={email}
+        onChange={onChange}
+        type="text"
+        placeholder="Email Address"
+      />
+      <button disabled={email === ''} type="submit">
+        Reset My Password
+      </button>
+    </form>
+  )
 }
 
-class PasswordForgetFormBase extends Component {
-  constructor(props) {
-    super(props)
-
-    this.state = { ...INITIAL_STATE }
-  }
-
-  onSubmit = event => {
-    const { email } = this.state
-
-    this.props.firebase
-      .doPasswordReset(email)
-      .then(() => {
-        this.setState({ ...INITIAL_STATE })
-      })
-      .catch(error => {
-        this.setState({ error })
-      })
-
-    event.preventDefault()
-  }
-
-  onChange = event => {
-    this.setState({ [event.target.name]: event.target.value })
-  }
-
-  render() {
-    const { email, error } = this.state
-
-    const isInvalid = email === ''
-
-    return (
-      <form onSubmit={this.onSubmit}>
-        <input
-          name="email"
-          value={this.state.email}
-          onChange={this.onChange}
-          type="text"
-          placeholder="Email Address"
-        />
-        <button disabled={isInvalid} type="submit">
-          Reset My Password
-        </button>
-
-        {error && <p>{error.message}</p>}
-      </form>
-    )
-  }
-}
-
-const PasswordForgetLink = () => (
-  <p>
-    <Link to={ROUTES.PASSWORD_FORGET}>Forgot Password?</Link>
-  </p>
-)
-
-export default PasswordForgetPage
-
-const PasswordForgetForm = withFirebase(PasswordForgetFormBase)
-
-export { PasswordForgetForm, PasswordForgetLink }
+export default withFirebase(PasswordForgetFormBase)
