+++ client/src/components/SignIn/index.js
@@ -3,7 +3,7 @@ import { withRouter } from 'react-router-dom'
 import { compose } from 'recompose'
 
 import { SignUpLink } from '../SignUp/SignUpLink'
-import { PasswordForgetLink } from '../PasswordForget'
+import PasswordForgetLink from '../PasswordForget'
 import { withFirebase } from '../../containers/Firebase'
 import { SignInFormBase } from './SignInFormBase'
 import { SignInGoogleBase } from './SignInGoogleBase'
