+++ client/src/constants/initialState.js
@@ -1,0 +1,4 @@
+export const INITIAL_STATE = {
+  email: '',
+  error: null
+}
