+++ client/src/containers/App.js
@@ -1,17 +1,17 @@
 import React from 'react'
 import { BrowserRouter as Router, Route } from 'react-router-dom'
 
-import Navigation from '../../views/Navigation'
-import LandingPage from '../../views/Landing'
-import SignUpPage from '../../components/SignUp/index'
-import SignInPage from '../../components/SignIn/index'
-import PasswordForgetPage from '../../components/PasswordForget/index'
-import HomePage from '../../views/Home'
-import AccountPage from '../../views/Account'
-import AdminPage from '../../views/Admin'
+import Navigation from '../views/Navigation'
+import LandingPage from '../views/Landing'
+import SignUpPage from '../components/SignUp'
+import SignInPage from '../components/SignIn'
+import PasswordForgetPage from '../components/PasswordForget'
+import HomePage from '../views/Home'
+import AccountPage from '../views/Account'
+import AdminPage from '../views/Admin'
 
-import * as ROUTES from '../../constants/routes'
-import { withAuthentication } from '../Sessions'
+import * as ROUTES from '../constants/routes'
+import { withAuthentication } from './Sessions'
 
 const App = () => (
   <Router>
