+++ client/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from './containers/App/App'
+import App from './containers/App'
 
 import Firebase, { FirebaseContext } from './containers/Firebase'
